tests/ciphex-presale.test.ts
import {
  assert,
  describe,
  test,
  clearStore,
  beforeAll,
  afterAll
} from "matchstick-as/assembly/index"
import { Address, BigInt } from "@graphprotocol/graph-ts"
import { Bought } from "../generated/schema"
import { Bought as BoughtEvent } from "../generated/CiphexPresale/CiphexPresale"
import { handleBought } from "../src/ciphex-presale"
import { createBoughtEvent } from "./ciphex-presale-utils"

// Tests structure (matchstick-as >=0.5.0)
// https://thegraph.com/docs/en/developer/matchstick/#tests-structure-0-5-0

describe("Describe entity assertions", () => {
  beforeAll(() => {
    let recipient = Address.fromString(
      "0x0000000000000000000000000000000000000001"
    )
    let referral = Address.fromString(
      "0x0000000000000000000000000000000000000001"
    )
    let token = Address.fromString("0x0000000000000000000000000000000000000001")
    let tokenAmount = BigInt.fromI32(234)
    let ciphexAmount = BigInt.fromI32(234)
    let ethPrice = BigInt.fromI32(234)
    let tokenPrice = BigInt.fromI32(234)
    let newBoughtEvent = createBoughtEvent(
      recipient,
      referral,
      token,
      tokenAmount,
      ciphexAmount,
      ethPrice,
      tokenPrice
    )
    handleBought(newBoughtEvent)
  })

  afterAll(() => {
    clearStore()
  })

  // For more test scenarios, see:
  // https://thegraph.com/docs/en/developer/matchstick/#write-a-unit-test

  test("Bought created and stored", () => {
    assert.entityCount("Bought", 1)

    // 0xa16081f360e3847006db660bae1c6d1b2e17ec2a is the default address used in newMockEvent() function
    assert.fieldEquals(
      "Bought",
      "0xa16081f360e3847006db660bae1c6d1b2e17ec2a-1",
      "recipient",
      "0x0000000000000000000000000000000000000001"
    )
    assert.fieldEquals(
      "Bought",
      "0xa16081f360e3847006db660bae1c6d1b2e17ec2a-1",
      "referral",
      "0x0000000000000000000000000000000000000001"
    )
    assert.fieldEquals(
      "Bought",
      "0xa16081f360e3847006db660bae1c6d1b2e17ec2a-1",
      "token",
      "0x0000000000000000000000000000000000000001"
    )
    assert.fieldEquals(
      "Bought",
      "0xa16081f360e3847006db660bae1c6d1b2e17ec2a-1",
      "tokenAmount",
      "234"
    )
    assert.fieldEquals(
      "Bought",
      "0xa16081f360e3847006db660bae1c6d1b2e17ec2a-1",
      "ciphexAmount",
      "234"
    )
    assert.fieldEquals(
      "Bought",
      "0xa16081f360e3847006db660bae1c6d1b2e17ec2a-1",
      "ethPrice",
      "234"
    )
    assert.fieldEquals(
      "Bought",
      "0xa16081f360e3847006db660bae1c6d1b2e17ec2a-1",
      "tokenPrice",
      "234"
    )

    // More assert options:
    // https://thegraph.com/docs/en/developer/matchstick/#asserts
  })
})
